feat: Expose the client connector publicly.
